+++ fidledemo/src/main/java/com/example/fidledemo/VO/GoodsItemVO.java
@@ -15,11 +15,11 @@ import java.util.List;
 @NoArgsConstructor
 public class GoodsItemVO {
 
-    private BigInteger id;
+    private Long id;
 
-    private BigInteger sellerId;
+    private Long sellerId;
 
-    private BigInteger buyerId;
+    private Long buyerId;
 
     private String title;
