+++ fidledemo/src/main/java/com/example/fidledemo/VO/TaskItemVO.java
@@ -15,11 +15,11 @@ import java.util.List;
 @NoArgsConstructor
 public class TaskItemVO {
 
-    private BigInteger id;
+    private Long id;
 
-    private BigInteger pulisherId;
+    private Long pulisherId;
 
-    private BigInteger accepterId;
+    private Long accepterId;
 
     private String title;
